// src/app/types.ts
// Global Types:

export type ViewPhase = 'select' | 'intro' | 'forces' | 'pulse' | 'thanks';

export interface PalaxyState {
    phase: ViewPhase;
    avatar?: Avatar;
    selectedAvatar?: Avatar;
    surveyResponses: Record<string, string>;
}

export interface ForceRelation {
    id: string;
    note: string;
}

export interface Forces {
    neighbors: ForceRelation[];
    counterforce: ForceRelation;
}

export interface Avatar {
    id: string;
    name: string;
    title: string;
    intro: string;
    bio: string;
    image: string;
    forces: Forces;
    species: string;
    homeworld: string;
    planet: string;
    synergy: string;
    psyche: string;
    wavelength: string;
}

export type AvatarBaseProps = {
    avatar: Avatar;
};

export type AvatarWithListProps = AvatarBaseProps & {
    mockAvatars: Avatar[];
};

export type ForceMapProps = AvatarWithListProps & {
    onNext: () => void;
};

export interface Question {
    id: number;
    text: string;
    options: string[];
}

export type FieldErrors<T> = Partial<Record<keyof T, string[]>>;